Add create products and the plumbing that go with
Add a create product rest api but that's not the interesting part, we
have now a reset of the tables between each test in tearDown so we are
not goign conflicts between tests.

There is some bit and pieces that need to be cleaned up but that give us
a base.

Change-Id: I0f14b9fa8c874c38f842a81fd8480962b273be70