Code refactoring
Full refactoring of the code, rewrite most parts to be more inline
between the db api and the rest api.

I have axed most of the part that we don't need yet and keep only the
base to be able to do the minimal calls of products (except for the
models).

A lot of the code is refactored from what we had before, what's inside
ironic (which is a more straightforward code) and some addded stuff by
me where it makese sense.

The testing is now doing a rollbacks after each test instead of trying to
delete the tables (like done in ceilometer).

Change-Id: I4e2c8ca2046695d8786c3089fa6ae522d97d4a30