Readd py3 support
Just needed an update keystonemiddleware which include my fix, there is
no releases with it so have to install from git.

For testing we have to remove the testrepository cause py27 testr and
py34 db format is not compatible

flake8 is broken it thinks that keystonemiddleware is a standard lib
cause we are doing a checkout in the requirements. Let make him happy for now until
keystonemiddleware is releasing a fix in pypi.

Signed-off-by: Chmouel Boudjnah <chmouel.boudjnah@enovance.com>
Change-Id: I2a139303a72e7cd5e80cc2202a451360f71fbc8e