Add ACL informations and context to requests
Add Request information about the ownership of a keystone token.
We don't need for api_paste to have keystone middleware anymore since we
are taking care of that part. We allow only certain parts of the API as
public and all others as protected down by auth_token.

This would be a build block to have ACL on methods

Change-Id: I76bc57d23b1d86e4eab274609a95bfcf8c5580be