Reintroduce MySQL
alow to mix the testing betwen SQLite and MySQL

This probably would have issues with parallized testing because we doing
some weird crap to get them rollbacked before each tests.

This works in production cause we just have a real world conflict there.

We can at leaast be able to do testing with MySQL with that.

Change-Id: Ie7418f349b08beb817a8746798bf102835b4afbd