Redo the functional tests in a nicer way
Rename to saladier_integrationtests

Use saladierclient library as backend for accesing the saladier.

Proper tests with all the plumbing that goes with it, which means we
have testr/testtools asserts and such integrations

Let it be able to be configured, so they can be used between different
saladier to validate it (i.e: a la tempest) in case you know if we want
to do CI or CD with the saladier.

Change-Id: Ifeef89d1ff4c36a99c325c47884888d6848d4cba